# src/multipart-upload/multipart-upload.controller.ts
import { Body, Controller, Delete, Get, Param, Patch, Post, Put } from '@nestjs/common';
import { MultipartUploadService } from './multipart-upload.service';
import { Authed } from '@/common/decorators/auth.decorator';
import {
  AbortMultipartUploadResponse,
  CompleteMultipartUploadResponse,
  ListMultipartUploadResponse,
  PostMultipartUploadResponse,
  StartMultipartUploadRequest,
  StartMultipartUploadResponse,
} from './multipart-upload.dto';
import { ApiBody, ApiConsumes, ApiOkResponse } from '@nestjs/swagger';
import { Permission } from '@/common/decorators/permission.decorator';
import { User } from '@/users/users.dto';
import { CsvValidationPipe } from '@/common/pipes/csv-validation.pipe';

@Controller('multipart-upload')
export class MultipartUploadController {
  constructor(private readonly multipartUploadService: MultipartUploadService) {}

  @Get()
  @ApiOkResponse({ type: ListMultipartUploadResponse })
  @Permission('list:multipart_upload')
  async listMultipartUploads(@Authed() user: User): Promise<ListMultipartUploadResponse> {
    return this.multipartUploadService.listMultipartUploads(user);
  }

  @Post()
  @ApiBody({ type: StartMultipartUploadRequest })
  @ApiOkResponse({ type: StartMultipartUploadResponse })
  @Permission('start:multipart_upload')
  async startMultipartUpload(
    @Authed() user: User,
    @Body() body: StartMultipartUploadRequest,
  ): Promise<StartMultipartUploadResponse> {
    return this.multipartUploadService.startMultipartUpload(user, body);
  }

  @Put(':uploadId')
  @ApiConsumes('text/csv')
  @ApiBody({ description: 'CSVデータ', type: String })
  @ApiOkResponse({ type: PostMultipartUploadResponse })
  @Permission('upload:multipart_upload')
  async uploadMultipartUpload(
    @Body(new CsvValidationPipe()) body: string,
    @Authed() user: User,
    @Param('uploadId') uploadId: string,
  ): Promise<PostMultipartUploadResponse> {
    return this.multipartUploadService.uploadMultipartUpload(user, uploadId, body);
  }

  @Patch(':uploadId')
  @ApiOkResponse({ type: CompleteMultipartUploadResponse })
  @Permission('complete:multipart_upload')
  async completeMultipartUpload(
    @Authed() user: User,
    @Param('uploadId') uploadId: string,
  ): Promise<CompleteMultipartUploadResponse> {
    return this.multipartUploadService.completeMultipartUpload(user, uploadId);
  }

  @Delete(':uploadId')
  @ApiOkResponse({ type: AbortMultipartUploadResponse })
  @Permission('abort:multipart_upload')
  async abortMultipartUpload(
    @Authed() user: User,
    @Param('uploadId') uploadId: string,
  ): Promise<AbortMultipartUploadResponse> {
    return this.multipartUploadService.abortMultipartUpload(user, uploadId);
  }
}