# src/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getRoot",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/hello": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "App"
        ]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/me": {
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetUserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/users/plan": {
      "get": {
        "operationId": "UsersController_getPlan",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/multipart-upload": {
      "get": {
        "operationId": "MultipartUploadController_listMultipartUploads",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMultipartUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "MultipartUpload"
        ]
      },
      "post": {
        "operationId": "MultipartUploadController_startMultipartUpload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMultipartUploadRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartMultipartUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "MultipartUpload"
        ]
      }
    },
    "/multipart-upload/{uploadId}": {
      "put": {
        "operationId": "MultipartUploadController_uploadMultipartUpload",
        "parameters": [
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "CSVデータ",
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostMultipartUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "MultipartUpload"
        ]
      },
      "patch": {
        "operationId": "MultipartUploadController_completeMultipartUpload",
        "parameters": [
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteMultipartUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "MultipartUpload"
        ]
      },
      "delete": {
        "operationId": "MultipartUploadController_abortMultipartUpload",
        "parameters": [
          {
            "name": "uploadId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AbortMultipartUploadResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "MultipartUpload"
        ]
      }
    },
    "/sensor-data": {
      "get": {
        "operationId": "SensorDataController_listSensorData",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "path",
            "description": "ページ番号",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "path",
            "description": "1ページあたりの件数",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSensorDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      },
      "post": {
        "operationId": "SensorDataController_uploadSensorDataFiles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/UploadSensorDataRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      }
    },
    "/sensor-data/{id}": {
      "get": {
        "operationId": "SensorDataController_getSensorData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSensorDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      },
      "patch": {
        "operationId": "SensorDataController_updateSensorData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDataRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSensorDataResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      },
      "delete": {
        "operationId": "SensorDataController_deleteSensorData",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      }
    },
    "/sensor-data/{id}/presigned-url": {
      "get": {
        "operationId": "SensorDataController_getSensorDataPresignedUrl",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetSensorDataPresignedUrlResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "SensorData"
        ]
      }
    },
    "/rate-limit/test": {
      "get": {
        "operationId": "RateLimitController_testRateLimit",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt": []
          }
        ],
        "tags": [
          "RateLimit"
        ]
      }
    }
  },
  "info": {
    "title": "Sensing Flow API",
    "description": "API documentation for Sensing Flow",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "jwt",
        "description": "Enter your JWT",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "ユーザー名"
          },
          "plan": {
            "type": "string",
            "enum": [
              "guest",
              "trial",
              "basic",
              "pro",
              "admin",
              "developer"
            ],
            "description": "プラン"
          }
        },
        "required": [
          "name",
          "plan"
        ]
      },
      "CreateUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ユーザーID"
          },
          "name": {
            "type": "string",
            "description": "ユーザー名"
          },
          "plan": {
            "type": "string",
            "enum": [
              "guest",
              "trial",
              "basic",
              "pro",
              "admin",
              "developer"
            ],
            "description": "プラン"
          }
        },
        "required": [
          "id",
          "name",
          "plan"
        ]
      },
      "GetUserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ユーザーID"
          },
          "name": {
            "type": "string",
            "description": "ユーザー名"
          },
          "plan": {
            "type": "string",
            "enum": [
              "guest",
              "trial",
              "basic",
              "pro",
              "admin",
              "developer"
            ],
            "description": "プラン"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "登録日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          }
        },
        "required": [
          "id",
          "name",
          "plan",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetPlanResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "type": "string",
            "enum": [
              "guest",
              "trial",
              "basic",
              "pro",
              "admin",
              "developer"
            ],
            "description": "プラン"
          }
        },
        "required": [
          "plan"
        ]
      },
      "MultiPartUpload": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "description": "センサデータのアップロードID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "status": {
            "type": "string",
            "description": "ステータス",
            "enum": [
              "in_progress",
              "completed",
              "aborted"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          }
        },
        "required": [
          "uploadId",
          "dataName",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListMultipartUploadResponse": {
        "type": "object",
        "properties": {
          "multipartUploads": {
            "description": "センサデータアップロード一覧",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MultiPartUpload"
            }
          }
        },
        "required": [
          "multipartUploads"
        ]
      },
      "StartMultipartUploadRequest": {
        "type": "object",
        "properties": {
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "dataName"
        ]
      },
      "StartMultipartUploadResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "description": "センサデータのアップロードID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "uploadId",
          "dataName"
        ]
      },
      "PostMultipartUploadResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "description": "センサデータのアップロードID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "uploadId",
          "dataName"
        ]
      },
      "CompleteMultipartUploadResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "description": "センサデータのアップロードID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "uploadId",
          "dataName"
        ]
      },
      "AbortMultipartUploadResponse": {
        "type": "object",
        "properties": {
          "uploadId": {
            "type": "string",
            "description": "センサデータのアップロードID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "uploadId",
          "dataName"
        ]
      },
      "SensorData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "センサデータID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          }
        },
        "required": [
          "id",
          "dataName",
          "createdAt",
          "updatedAt"
        ]
      },
      "ListSensorDataResponse": {
        "type": "object",
        "properties": {
          "sensorData": {
            "description": "センサデータ一覧",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorData"
            }
          }
        },
        "required": [
          "sensorData"
        ]
      },
      "UploadSensorDataRequest": {
        "type": "object",
        "properties": {
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          }
        },
        "required": [
          "dataName",
          "createdAt"
        ]
      },
      "GetSensorDataResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "センサデータID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          }
        },
        "required": [
          "id",
          "dataName",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateSensorDataRequest": {
        "type": "object",
        "properties": {
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          }
        },
        "required": [
          "dataName"
        ]
      },
      "UpdateSensorDataResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "センサデータID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          }
        },
        "required": [
          "id",
          "dataName",
          "createdAt",
          "updatedAt"
        ]
      },
      "GetSensorDataPresignedUrlResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "センサデータID"
          },
          "dataName": {
            "type": "string",
            "description": "センサデータ名"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "作成日時"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "更新日時"
          },
          "presignedUrl": {
            "type": "string",
            "description": "presigned URL"
          }
        },
        "required": [
          "id",
          "dataName",
          "createdAt",
          "updatedAt",
          "presignedUrl"
        ]
      }
    }
  }
}